게시글 조회 page 다시 처리

+++ src/main/java/com/lth/community/vo/board/GetBoardVO.java
@@ -1,17 +1,9 @@
 package com.lth.community.vo.board;
 
-import com.lth.community.entity.BoardInfoEntity;
-import com.lth.community.entity.MemberInfoEntity;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
@@ -20,7 +12,7 @@ import java.util.List;
 @Builder
 public class GetBoardVO {
     private List<BoardInfoVO> list;
-    private Long total;
+    private Integer total;
     private Integer totalPage;
     private Integer currentPage;
 }
